Fix bug in temperature sensor calibration
This patch fixes a floating-point rounding issue
in the calibration function of the temperature sensor.
It ensures accurate readings for high-precision devices.
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch JiraHCMUT#11_Fix_timer_of_light
Your branch is up to date with 'origin/JiraHCMUT#11_Fix_timer_of_light'.

Changes to be committed:
      modified:   Core/Src/button.c

Change-Id: Ib420b3781f2c476df6a601bb6c22df03d29bb83c